pyproject.toml
[tool.poetry]
name = "database-hash-bcrypt"
version = "0.1.0"
description = ""
authors = ["brenoAV"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
psycopg2-binary = "^2.9.9"
bcrypt = "^4.1.2"


[tool.poetry.group.test.dependencies]
pytest = "^8.0.2"
pytest-mock = "^3.12.0"


[tool.poetry.group.dev.dependencies]
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
